Add options to rotate a row or column.
This release contains a bug!
See the new TODO comment.
The bug is in code that hasn't changed.
The new code helped me find the old bug.

Set the background color of each cell to match the text.
This is temporary to help debug before the real GUI is finished.