Bug fix
The program was crashing sometimes.
See the comments in the diff for more details about the bug.

Added a lot of debug code.

Added a white border to some of the text.
So the black letters will be easier to read on various backgrounds.
This text is all temporary until there is a better GUI.