Minor improvement to the GUI.
If you make a group of exactly 5 tiles, one of them will turn into a bomb.
The animation for that has changed slightly.
The tile will show both the group decoration and the bomb.
The bomb is partially covered by the decoration, but is still visible.

And minor code cleanup.