Interim release.
After a lot of reorganizing, not too much is broken.
I need to figure out why I had to add those two debug pauses.
And fix that problem.

With the reorganizing complete I will be able to animate the bombs when you get 6 tiles in one group.

Change the cell phone warning.
It looks like I fixed that problem a while ago.
I probably fixed it, but didn't realize it because of a testing problem.
It still needs more testing, but it looks good.