Minor bug fixes.
The updates now happen at full speed again.
I added some temporary delays to work around the bug.
Now wait for all the right promises.
I was approximating that with a timer, but that often failed.

The functionality is back to where it was several commits ago.
But the code is better organized so I can add more bomb stuff.