Remove a decoration that didn't work on my Android phone.
And lots of automatic reformatting.

Add white-space: nowrap to the decoration tester.
This was not an issue in Chrome on my Mac.
But my Android phone was wrapping more than it should have.
I'm trying to use the border to tell me what the decoration is doing to the margins.
If the div becomes two lines, the border doesn't tell me as much.