Finish the bomb throwing dance.
If you make a group of 6 or more pieces, each piece gets a bomb.
And the piece throws the bomb onto another piece, before the original piece goes flying off the board.

I want to do some minor cleanup and documentation before calling it done.
I'll do a build with the next commit.
I'm freezing this now because the code is working.