Add artwork for an explosion.
Currently it is only visible through the debugger.
It can be displayed via the explosionVisible property.
There is already code counting the number of items affected by the bomb.
I am not yet removing the items that have been bombed.
The next TODO is to show the explosions and remove the squares, when bombed.